fix: suppr livraison si intégration terminée #765

+++ src/Controller/Entrepot/UploadController.php
@@ -146,7 +146,7 @@ class UploadController extends AbstractController implements ApiControllerInterf
             }
 
             // supprime livraison si intégration terminée
-            if ($uploadIntegrationWorkflow->isIntegrationCompleted($progress)) {
+            if (false === $getOnlyProgress && $uploadIntegrationWorkflow->isIntegrationCompleted($progress)) {
                 $this->uploadApiService->remove($datastoreId, $uploadId);
             }
